requirement.txt
django==5.0.6
django-bootstrap5
django-crispy-forms
crispy-bootstrap5
django-phonenumber-field[phonenumbers]
requests
git+https://github.com/azure-samples/ms-identity-python-utilities@main